feat - 누락된 부분 (#77)
* feat(Setup): Initial setup Page settings

* feat(PolicyAgreement):  로고 배치

* feat(PolicyDocumentSheet): UI 개선

* feat(NameSetup): UI 개선

* feat(Name,Tag): 이름, 태그 UI 및 텍스트 수정

* feat(InitialSetupFlow): 건너뛰기 버튼 UI 통일

* feat(Setup): 튜토리얼 페이지 UI/UX 최종 업데이트

* feat(AddSearchDetail): 포커스 상태해제 UX 업데이트

* feat(Login): 로고 위치 조정

* refactor(Setup): 컴포넌트 분리

* feat(1.1.8): 버전 업데이트
