+++ KillingPart/Views/Screens/Setup/components/NameSetupScreen.swift
@@ -1,0 +1,83 @@
+import SwiftUI
+
+struct NameSetupScreen: View {
+    @ObservedObject var viewModel: InitialSetupFlowViewModel
+
+    var body: some View {
+        GeometryReader { geometry in
+            let horizontalPadding = AppSpacing.l
+            let safeAreaBottomInset = geometry.safeAreaInsets.bottom
+            let isKeyboardPresented = safeAreaBottomInset > 80
+            let bottomPadding = isKeyboardPresented
+                ? AppSpacing.s
+                : safeAreaBottomInset + AppSpacing.l
+            let contentTopOffset = max(
+                geometry.safeAreaInsets.top + AppSpacing.xl,
+                geometry.size.height * 0.22
+            )
+
+            VStack(spacing: 0) {
+                VStack(alignment: .leading, spacing: AppSpacing.s) {
+                    Text("사용하실 이름이 무엇인가요?")
+                        .font(AppFont.paperlogy7Bold(size: 24))
+                        .foregroundStyle(.white)
+
+                    Text("이름은 언제든지 바꿀 수 있습니다")
+                        .font(AppFont.paperlogy4Regular(size: 14))
+                        .foregroundStyle(.white)
+
+                    Text("*한글 8자 이내, 영어 16자 이내(숫자 포함)")
+                        .font(AppFont.paperlogy4Regular(size: 12))
+                        .foregroundStyle(.white.opacity(0.55))
+
+                    TextField("이름 입력", text: $viewModel.nameDraft)
+                        .font(AppFont.paperlogy5Medium(size: 16))
+                        .foregroundStyle(.white)
+                        .textInputAutocapitalization(.never)
+                        .autocorrectionDisabled()
+                        .padding(.horizontal, AppSpacing.m)
+                        .padding(.vertical, AppSpacing.m)
+                        .background(Color.white.opacity(0.08))
+                        .clipShape(RoundedRectangle(cornerRadius: 18, style: .continuous))
+                        .overlay {
+                            RoundedRectangle(cornerRadius: 18, style: .continuous)
+                                .stroke(Color.white.opacity(0.1), lineWidth: 1)
+                        }
+
+                    if let validation = viewModel.validateName(viewModel.nameDraft), !viewModel.nameDraft.isEmpty {
+                        Text(validation)
+                            .font(AppFont.paperlogy4Regular(size: 13))
+                            .foregroundStyle(.red.opacity(0.95))
+                    } else if let errorMessage = viewModel.errorMessage {
+                        Text(errorMessage)
+                            .font(AppFont.paperlogy4Regular(size: 13))
+                            .foregroundStyle(.red.opacity(0.95))
+                    }
+                }
+                .frame(maxWidth: .infinity, alignment: .leading)
+                .padding(.horizontal, horizontalPadding)
+                .padding(.top, contentTopOffset)
+
+                Spacer(minLength: 0)
+
+                PrimaryButton(title: "다음", isLoading: viewModel.isLoading) {
+                    Task {
+                        await viewModel.submitName()
+                    }
+                }
+                .disabled(!viewModel.canSubmitName || viewModel.isLoading)
+                .opacity((viewModel.canSubmitName && !viewModel.isLoading) ? 1 : 0.45)
+                .padding(.horizontal, horizontalPadding)
+                .padding(.bottom, bottomPadding)
+            }
+        }
+        .background(
+            LinearGradient(
+                colors: [Color.black, Color(hex: "#10131B")],
+                startPoint: .topLeading,
+                endPoint: .bottomTrailing
+            )
+            .ignoresSafeArea()
+        )
+    }
+}
